refactor(slice-45): extract FE screen chrome, hooks, and explore panels
Shrink god screens via presentational siblings and useExperimentDetail
before the components folder move, keeping dashboard behaviour unchanged.
