+++ CHANGELOG.md
@@ -52,6 +52,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - **Slice 45 (rerank package)** — moved `reranker` + `local_reranker` into `server/core/rerank/`; shim aliases keep old import/patch paths working.
 - **Slice 45 (db theme packages)** — moved ports/factory into `server/db/ports/`, Mongo adapters into `server/db/mongo/`, Postgres adapters + `schema.sql` into `server/db/postgres/`; shim aliases at old `server.db.*` paths.
 - **Slice 45 (fat-surface slim)** — extracted Bayesian/stale-status helpers to `server/api/experiments_lifecycle.py` (`experiments.py` 553→388) and CLI watch/summary presentation to `cli/display.py` (`cli/main.py` 519→246).
+- **Slice 45 (FE screen SLAP)** — extracted `hooks/useExperimentDetail`, `components/experimentDetail/*`, `experimentList/labels`, `explore/ExplorePanels` (+ duration helpers on `experimentDetailProgress`); screen line counts Detail 1615→986, Experiments 764→670, Explorer 1131→426 (DECISIONS #156).
 - **Slice 45 (FE shared primitives)** — extracted `components/chrome/Pagination`, `components/stats/StatTile`+`StatRow`, `utils/feedEntries.appendFeedEntry`, `utils/completionReason.completionReasonLabel`; screens/stats panels compose them (DECISIONS #155).
 - **Slice 45 (tests mirror + mega-suite split)** — flat `tests/test_*.py` mirrored under `tests/server/{core,db,api,models}/`, `tests/cli/`, `tests/scripts/`; `test_slice16_parallel_sweep.py` (~2296) split into focused `tests/server/core/pipeline/` modules + `tests/helpers/pipeline_sweep.py`.
 - **Nightly Meterian OSS path** — `nightly.yml` Meterian job uses `oss: true` (no `METERIAN_API_TOKEN`), pins scanners to Python + Node (`cli_args`), and archives `meterian-<run>` artifacts (HTML, JUnit, SARIF, CycloneDX + CSV SBOM) for comparison with Anchore (**IMPLEMENTED**; not yet observed on a green nightly run).
