+++ frontend/src/components/ExperimentDetailScreen.tsx
@@ -6,35 +6,30 @@
  * - Themed metadata sections (Git, Config, Data, Sweep Dimensions)
  * - Enhanced runs table with color-coded badges
  * - Success/failure panels with large visual feedback
- *
- * Design Decisions:
- * - "Embedding Models" (not "Models") to distinguish from reranking models
- * - Voyage rate limit fields only shown when provider is "voyage"
- * - Failed count always visible (even when 0) for consistency
- * - Color system: Blue (primary), Green (success), Red (failure), Amber (warning), Purple (secondary)
  */
-import { useCallback, useEffect, useRef, useState } from 'react';
-import { DETAIL_POLL_MS, DEV_POLL_LOG_INTERVAL_MS, LOADING_STALL_AFTER_MS, LOADING_STALL_REPEAT_MS, VECTOR_DB_STATS_POLL_MS } from '../constants';
+import { DETAIL_POLL_MS } from '../constants';
 import AppPageChrome from './AppPageChrome';
 import DashboardShell from './DashboardShell';
 import LoadingFeedbackPanel from './LoadingFeedbackPanel';
 import ExperimentProgressCard from './ExperimentProgressCard';
 import ExperimentVectorDbStatsCard from './ExperimentVectorDbStatsCard';
-import type { FeedEntry } from './LoadingFeedbackPanel';
-import {
-  deleteExperiment,
-  getExperiment,
-  getExperimentDbStats,
-  getExperimentWithProgress,
-  type ExperimentProgressCallback,
-} from '../services/apiClient';
 import ExperimentControlButtons from './ExperimentControlButtons';
 import ConfirmDeleteModal from './ConfirmDeleteModal';
 import CollapsibleCard from './CollapsibleCard';
-import { RunStatus, Phase, EnvParams, SweepSummary, ExperimentStatus, Experiment, ExperimentDbStatsSummary } from '../types';
-import { createStallWatcher, type FetchProgressUpdate } from '../services/fetchWithProgress';
-import { devInfo, devInfoThrottled, devWarn } from '../utils/devLog';
-import { toExperimentDbStatsSummary } from '../utils/experimentDbStats';
+import Pagination from './chrome/Pagination';
+import {
+  detailIcons as icons,
+  DimensionBadge,
+  MetadataItem,
+  PhaseIndicator,
+  ProgressSubtitle,
+  StatCard,
+  StatusBadge,
+} from './experimentDetail';
+import { formatDurationFromRuns } from './experimentDetailProgress';
+import { useExperimentDetail } from '../hooks/useExperimentDetail';
+import { Phase, Experiment } from '../types';
+import type { ExperimentDbStatsSummary } from '../types';
 import {
   displayDatabaseProvider,
   experimentHydratingBlurb,
@@ -46,392 +41,7 @@ import {
   isTerminalExperimentStatus,
   summarizeExperimentRuns,
 } from '../utils/experimentStatus';
-import {
-  calculateProgressMetrics,
-  formatTimeWithUnits,
-} from './experimentDetailProgress';
-import Pagination from './chrome/Pagination';
-import { appendFeedEntry } from '../utils/feedEntries';
 import { completionReasonLabel } from '../utils/completionReason';
-
-// Icon components (minimal SVG)
-const icons = {
-  clock: (
-    <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
-      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z" />
-    </svg>
-  ),
-  check: (
-    <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
-      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M5 13l4 4L19 7" />
-    </svg>
-  ),
-  x: (
-    <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
-      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M6 18L18 6M6 6l12 12" />
-    </svg>
-  ),
-  play: (
-    <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
-      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M14.752 11.168l-3.197-2.132A1 1 0 0010 9.87v4.263a1 1 0 001.555.832l3.197-2.132a1 1 0 000-1.664z" />
-      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 12a9 9 0 11-18 0 9 9 0 0118 0z" />
-    </svg>
-  ),
-  code: (
-    <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
-      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M10 20l4-16m4 4l4 4-4 4M6 16l-4-4 4-4" />
-    </svg>
-  ),
-  database: (
-    <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
-      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 7v10c0 2.21 3.582 4 8 4s8-1.79 8-4V7M4 7c0 2.21 3.582 4 8 4s8-1.79 8-4M4 7c0-2.21 3.582-4 8-4s8 1.79 8 4m0 5c0 2.21-3.582 4-8 4s-8-1.79-8-4" />
-    </svg>
-  ),
-  settings: (
-    <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
-      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M10.325 4.317c.426-1.756 2.924-1.756 3.35 0a1.724 1.724 0 002.573 1.066c1.543-.94 3.31.826 2.37 2.37a1.724 1.724 0 001.065 2.572c1.756.426 1.756 2.924 0 3.35a1.724 1.724 0 00-1.066 2.573c.94 1.543-.826 3.31-2.37 2.37a1.724 1.724 0 00-2.572 1.065c-.426 1.756-2.924 1.756-3.35 0a1.724 1.724 0 00-2.573-1.066c-1.543.94-3.31-.826-2.37-2.37a1.724 1.724 0 00-1.065-2.572c-1.756-.426-1.756-2.924 0-3.35a1.724 1.724 0 001.066-2.573c-.94-1.543.826-3.31 2.37-2.37.996.608 2.296.07 2.572-1.065z" />
-      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 12a3 3 0 11-6 0 3 3 0 016 0z" />
-    </svg>
-  ),
-  grid: (
-    <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
-      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 5a1 1 0 011-1h4a1 1 0 011 1v4a1 1 0 01-1 1H5a1 1 0 01-1-1V5zM14 5a1 1 0 011-1h4a1 1 0 011 1v4a1 1 0 01-1 1h-4a1 1 0 01-1-1V5zM4 15a1 1 0 011-1h4a1 1 0 011 1v4a1 1 0 01-1 1H5a1 1 0 01-1-1v-4zM14 15a1 1 0 011-1h4a1 1 0 011 1v4a1 1 0 01-1 1h-4a1 1 0 01-1-1v-4z" />
-    </svg>
-  ),
-  search: (
-    <svg className="h-4 w-4 shrink-0" fill="none" viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
-      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="m21 21-4.35-4.35m1.35-5.65a7 7 0 11-14 0 7 7 0 0114 0z" />
-    </svg>
-  ),
-  trash: (
-    <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
-      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 7h16m-10 4v6m4-6v6m-7-10 .75 13h8.5L17 7m-7-3h4l1 3H9l1-3z" />
-    </svg>
-  ),
-  document: (
-    <svg className="h-4 w-4 shrink-0" fill="none" viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
-      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M7 3h7l4 4v14H7V3zm7 0v5h5M10 12h5m-5 4h5" />
-    </svg>
-  ),
-  back: (
-    <svg className="h-4 w-4" fill="none" viewBox="0 0 24 24" stroke="currentColor" aria-hidden="true">
-      <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="m15 18-6-6 6-6" />
-    </svg>
-  ),
-};
-
-interface ExperimentDetail {
-  experiment_id: string;
-  experiment_name: string;
-  status: ExperimentStatus;
-  created_at?: string;
-  completed_at?: string | null;
-  run_count?: number;
-  grid_equivalent_count?: number;
-  failed_count?: number;
-  completion_reason?: string;
-  bayesian_summary?: {
-    best_query_avg_score?: number;
-    best_chunk_size?: number;
-    best_overlap?: number;
-    best_embedding_model?: string;
-    best_retrieval_method?: string;
-    best_retriever_type?: string;
-    grid_equivalent_count?: number;
-    planned_trials?: number;
-    attempted_trials?: number;
-    discarded_trials?: number;
-    not_started?: number;
-    termination_reason?: string;
-  };
-  runs?: RunStatus[];
-  started_at?: string;
-  git_commit?: string;
-  git_branch?: string;
-  git_dirty?: boolean;
-  python_version?: string;
-  app_version?: string;
-  env_params?: EnvParams;
-  data_paths?: string[];
-  queries_file?: string;
-  retrieval_model?: string | null;
-  top_k_initial?: number;
-  top_k_final?: number;
-  parallelism?: number;
-  on_error?: string;
-  sweep_summary?: SweepSummary;
-  config?: {
-    embedding?: {
-      provider?: string;
-    };
-    retrieval?: {
-      retrieval_provider?: string;
-    };
-    execution?: {
-      search_strategy?: 'grid' | 'bayesian';
-    };
-  };
-}
-
-const PHASE_ORDER: Phase[] = [
-  Phase.QUEUED, Phase.PARSING, Phase.CHUNKING, Phase.EMBEDDING,
-  Phase.STORING, Phase.QUERYING, Phase.RERANKING, Phase.COMPLETE,
-];
-
-function PhaseIndicator({ current }: { current: Phase }) {
-  const currentIdx = PHASE_ORDER.indexOf(current);
-  const isFailed = current === Phase.FAILED || current === Phase.INTERRUPTED;
-  const safeCurrent = typeof current === 'string' ? current : 'unknown';
-
-  return (
-    <div className="relative group flex gap-1 items-center">
-      {PHASE_ORDER.map((phase, i) => {
-        const isCurrent = phase === current;
-        const isPast = i < currentIdx;
-
-        let bg = 'bg-slate-200';
-        if (isFailed) bg = 'bg-red-300';
-        else if (isCurrent) bg = 'bg-blue-500 animate-pulse';
-        else if (isPast) bg = 'bg-green-400';
-
-        return (
-          <div
-            key={phase}
-            className={`w-3 h-3 rounded-full ${bg}`}
-          />
-        );
-      })}
-      <span className="ml-2 text-xs text-slate-500">{safeCurrent}</span>
-
-      {/* Tooltip on hover */}
-      <div className="absolute left-0 bottom-full mb-2 hidden group-hover:block z-50 pointer-events-none">
-        <div className="bg-slate-800 text-white text-xs rounded-lg shadow-lg px-3 py-2 whitespace-nowrap">
-          {PHASE_ORDER.map((p, i) => {
-            let icon = '○';
-            let color = 'text-slate-400';
-            let label = '';
-            if (isFailed) {
-              icon = '✗';
-              color = 'text-red-300';
-            } else if (i < currentIdx) {
-              icon = '✓';
-              color = 'text-green-300';
-            } else if (i === currentIdx) {
-              icon = '●';
-              color = 'text-blue-300';
-              label = ' ← current';
-            }
-
-            return (
-              <div key={p} className={`${color} leading-5`}>
-                {icon} {p}{label}
-              </div>
-            );
-          })}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function formatDuration(startedAt?: string, completedAt?: string | null): string {
-  if (!startedAt || !completedAt) return '—';
-  const ms = new Date(completedAt).getTime() - new Date(startedAt).getTime();
-  if (ms < 1000) return `${ms}ms`;
-  const totalSeconds = ms / 1000;
-  return formatTimeWithUnits(totalSeconds);
-}
-
-function parseSafeTimestamp(value: string | undefined): number | null {
-  if (!value) return null;
-  const parsed = new Date(value).getTime();
-  if (Number.isNaN(parsed)) return null;
-  return parsed;
-}
-
-function formatDurationFromRuns(runs: RunStatus[] = [], startedAt?: string, completedAt?: string | null): string {
-  const runStartedAt = Math.min(
-    ...runs.map((run) => parseSafeTimestamp(run.created_at)).filter((value): value is number => value !== null),
-  );
-  const runCompletedAt = Math.max(
-    ...runs.map((run) => parseSafeTimestamp(run.updated_at)).filter((value): value is number => value !== null),
-  );
-
-  const allRunsHaveTimestamps =
-    Number.isFinite(runStartedAt) && Number.isFinite(runCompletedAt) && runCompletedAt >= runStartedAt;
-
-  if (allRunsHaveTimestamps) {
-    return formatDuration(new Date(runStartedAt).toISOString(), new Date(runCompletedAt).toISOString());
-  }
-
-  return formatDuration(startedAt, completedAt);
-}
-
-function ProgressSubtitle({
-  completed,
-  total,
-  startedAt,
-}: {
-  completed: number;
-  total: number;
-  startedAt?: string;
-}) {
-  const { elapsedStr, etaStr } = calculateProgressMetrics({
-    completed,
-    total,
-    startedAt,
-    now: new Date().getTime(),
-  });
-
-  return (
-    <div className="flex flex-wrap items-center gap-3 text-sm">
-      <span className="font-medium text-ink">
-        {completed} of {total} runs completed
-      </span>
-      <span className="text-line">•</span>
-      <span className="inline-flex items-center gap-1.5">
-        <span className="text-xs font-semibold uppercase tracking-wide text-cobalt">Elapsed</span>
-        <span className="font-mono font-semibold text-cobalt">{elapsedStr}</span>
-      </span>
-      <span className="text-line">•</span>
-      <span className="inline-flex items-center gap-1.5">
-        <span className="text-xs font-semibold uppercase tracking-wide text-accent-strong">ETA</span>
-        <span className="font-mono font-semibold text-accent-strong">{etaStr}</span>
-      </span>
-    </div>
-  );
-}
-
-function MetadataItem({ label, value }: { label: string; value: string | number | boolean | undefined }) {
-  if (value === undefined || value === null) return null;
-
-  // Format large numbers for readability
-  const formatValue = (val: string | number | boolean): string => {
-    if (typeof val === 'number') {
-      // Format TPM/RPM with commas for readability
-      if (label.includes('TPM') || label.includes('RPM')) {
-        return val.toLocaleString();
-      }
-      return String(val);
-    }
-    return String(val);
-  };
-
-  return (
-    <div className="flex flex-col">
-      <span className="text-xs uppercase tracking-wider text-muted">{label}</span>
-      <span className="font-mono text-sm text-ink">{formatValue(value)}</span>
-    </div>
-  );
-}
-
-// Status badge with color coding
-function StatusBadge({ status }: { status: string }) {
-  const configByStatus = {
-    complete: { bg: 'bg-emerald-50', text: 'text-emerald-800', icon: icons.check, border: 'border-emerald-200' },
-    running: { bg: 'bg-blue-50', text: 'text-blue-800', icon: icons.play, border: 'border-blue-200' },
-    failed: { bg: 'bg-red-50', text: 'text-red-800', icon: icons.x, border: 'border-red-200' },
-    partial: { bg: 'bg-amber-50', text: 'text-amber-900', icon: icons.x, border: 'border-amber-200' },
-    cancelled: { bg: 'bg-slate-100', text: 'text-slate-800', icon: icons.x, border: 'border-slate-300' },
-    paused: { bg: 'bg-violet-50', text: 'text-violet-800', icon: icons.clock, border: 'border-violet-200' },
-  } as const;
-  const defaultConfig = { bg: 'bg-canvas', text: 'text-ink', icon: icons.clock, border: 'border-line' };
-  const config =
-    status in configByStatus
-      ? configByStatus[status as keyof typeof configByStatus]
-      : defaultConfig;
-
-  return (
-    <div className={`inline-flex min-h-9 items-center gap-2 rounded-full border px-4 py-2 font-semibold ${config.bg} ${config.text} ${config.border}`}>
-      {config.icon}
-      <span className="uppercase text-sm tracking-wide">{status}</span>
-    </div>
-  );
-}
-
-// Stat card for key metrics
-function StatCard({
-  label,
-  value,
-  icon,
-  trend,
-  color = 'blue',
-  compact = false,
-}: {
-  label: string;
-  value: string | number;
-  icon: React.ReactNode;
-  trend?: string;
-  color?: 'blue' | 'green' | 'purple' | 'amber' | 'red' | 'slate';
-  compact?: boolean;
-}) {
-  const colors = {
-    blue: 'bg-blue-50 text-blue-600 border-blue-200',
-    green: 'bg-green-50 text-green-700 border-green-200',
-    purple: 'bg-accent-soft text-accent-strong border-accent',
-    amber: 'bg-amber-50 text-amber-700 border-amber-200',
-    red: 'bg-red-50 text-red-700 border-red-200',
-    slate: 'bg-slate-50 text-slate-600 border-slate-200',
-  } as const;
-  // color is a typed union — safe lookup, not user-controlled injection.
-  // eslint-disable-next-line security/detect-object-injection -- keyed by StatCard color prop union
-  const colorClass = colors[color];
-
-  if (compact) {
-    return (
-      <div className={`${colorClass} rounded-lg border px-3 py-2.5 min-w-0 h-full`}>
-        <div className="flex items-center gap-2 min-w-0">
-          <div className="shrink-0 scale-90 opacity-80">{icon}</div>
-          <div className="min-w-0 flex-1">
-            <div className="text-lg font-bold leading-none tabular-nums truncate">{value}</div>
-            <div className="mt-1 truncate text-xs font-semibold uppercase tracking-wide opacity-75">
-              {label}
-            </div>
-          </div>
-          {trend && (
-            <span className="shrink-0 text-xs font-medium opacity-75">{trend}</span>
-          )}
-        </div>
-      </div>
-    );
-  }
-
-  return (
-    <div className={`${colorClass} rounded-xl p-4 border-2 shadow-sm`}>
-      <div className="flex items-start justify-between mb-2">
-        <div className="p-2 rounded-lg bg-white/80">
-          {icon}
-        </div>
-        {trend && <span className="text-xs font-medium opacity-75">{trend}</span>}
-      </div>
-      <div className="mt-2">
-        <div className="text-2xl font-bold">{value}</div>
-        <div className="text-xs font-medium uppercase tracking-wide mt-1 opacity-75">{label}</div>
-      </div>
-    </div>
-  );
-}
-
-// Progress ring moved to ExperimentProgressCard.tsx (reusable component)
-
-// Dimension badge for sweep params
-function DimensionBadge({ label, values }: { label: string; values: (string | number)[] }) {
-  return (
-    <div className="rounded-xl border border-line bg-paper p-3">
-      <div className="mb-2 flex items-center gap-2">
-        <div className="h-2 w-2 rounded-full bg-accent"></div>
-        <span className="text-xs font-bold uppercase tracking-wider text-muted">{label}</span>
-      </div>
-      <div className="flex flex-wrap gap-1">
-        {values.map((v, i) => (
-          <span key={i} className="inline-flex items-center rounded-md border border-line bg-canvas px-2 py-1 text-xs font-medium text-ink">
-            {String(v)}
-          </span>
-        ))}
-      </div>
-    </div>
-  );
-}
 
 export default function ExperimentDetailScreen({
   experimentId,
@@ -446,270 +56,31 @@ export default function ExperimentDetailScreen({
   onBack: () => void;
   onExplore?: () => void;
 }) {
-  const seededDetail =
-    initialExperiment?.experiment_id === experimentId
-      ? (initialExperiment as unknown as ExperimentDetail)
-      : null;
-
-  const [detail, setDetail] = useState<ExperimentDetail | null>(seededDetail);
-  const [error, setError] = useState<string | null>(null);
-  const [hydrating, setHydrating] = useState(seededDetail === null);
-  const [showDeleteModal, setShowDeleteModal] = useState(false);
-  const [deleting, setDeleting] = useState(false);
-
-  const [loadFeed, setLoadFeed] = useState<FeedEntry[]>([]);
-  const [receivedBytes, setReceivedBytes] = useState<number | null>(null);
-  const [totalBytes, setTotalBytes] = useState<number | null>(null);
-
-  const [dbStats, setDbStats] = useState<ExperimentDbStatsSummary | null>(initialDbStats ?? null);
-  const [dbStatsLoading, setDbStatsLoading] = useState(initialDbStats === undefined);
-
-  const [runsCurrentPage, setRunsCurrentPage] = useState(1);
-  const [runsItemsPerPage, setRunsItemsPerPage] = useState(15);
-
-  const handleRunsItemsPerPageChange = useCallback((items: number) => {
-    setRunsItemsPerPage(items);
-    setRunsCurrentPage(1);
-  }, []);
-
-  const aliveRef = useRef(true);
-  const pollRef = useRef<ReturnType<typeof setInterval> | null>(null);
-  const pollDevLogAtRef = useRef(new Map<string, number>());
-  const dbStatsInFlightRef = useRef<Promise<void> | null>(null);
-
-  const experimentMeta = useCallback((): Pick<Experiment, 'experiment_id' | 'experiment_name' | 'status' | 'created_at'> | null => {
-    if (detail) {
-      return {
-        experiment_id: detail.experiment_id,
-        experiment_name: detail.experiment_name,
-        status: detail.status,
-        created_at: detail.created_at ?? initialExperiment?.created_at ?? new Date(0).toISOString(),
-      };
-    }
-    if (initialExperiment?.experiment_id === experimentId) return initialExperiment;
-    return null;
-  }, [detail, initialExperiment, experimentId]);
-
-  const loadDbStats = useCallback(
-    async (options?: { showLoading?: boolean }) => {
-      if (dbStatsInFlightRef.current !== null) {
-        return dbStatsInFlightRef.current;
-      }
-
-      const request = (async () => {
-        const meta = experimentMeta();
-        if (!meta) return;
-        if (options?.showLoading) setDbStatsLoading(true);
-        try {
-          const response = await getExperimentDbStats(experimentId);
-          setDbStats(toExperimentDbStatsSummary(meta, response.db_stats));
-        } catch (err) {
-          devWarn('ExperimentDetailScreen', `db stats load failed — ${experimentId.slice(0, 8)}…`, err);
-        } finally {
-          dbStatsInFlightRef.current = null;
-          setDbStatsLoading(false);
-        }
-      })();
-
-      dbStatsInFlightRef.current = request;
-      return request;
-    },
-    [experimentId, experimentMeta],
-  );
-
-  useEffect(() => {
-    if (!initialDbStats) {
-      void loadDbStats({ showLoading: true });
-    }
-
-    const statsTimer = window.setInterval(() => {
-      void loadDbStats();
-    }, VECTOR_DB_STATS_POLL_MS);
-
-    return () => window.clearInterval(statsTimer);
-  }, [experimentId, initialDbStats, loadDbStats]);
-
-  const stopDetailPoll = useCallback(() => {
-    if (pollRef.current !== null) {
-      clearInterval(pollRef.current);
-      pollRef.current = null;
-    }
-  }, []);
-
-  const startDetailPollIfRunning = useCallback(
-    (status: ExperimentStatus | undefined, completedAt?: string | null) => {
-      stopDetailPoll();
-      if (!isRunningExperimentStatus(status, completedAt)) return;
-
-      devInfo('ExperimentDetailScreen', `detail poll started — ${experimentId.slice(0, 8)}… every ${DETAIL_POLL_MS}ms`);
-
-      pollRef.current = window.setInterval(async () => {
-        if (!aliveRef.current) return;
-        try {
-          const next = await getExperiment(experimentId);
-          if (!aliveRef.current) return;
-          setDetail(next as unknown as ExperimentDetail);
-          setError(null);
-          devInfoThrottled(
-            'ExperimentDetailScreen',
-            `poll:detail:${experimentId}`,
-            DEV_POLL_LOG_INTERVAL_MS,
-            `detail poll OK — ${experimentId.slice(0, 8)}… status=${next.status}`,
-            pollDevLogAtRef.current,
-          );
-          if (
-            isTerminalExperimentStatus(next.status, next.completed_at)
-          ) {
-            stopDetailPoll();
-          }
-        } catch (pollErr) {
-          if (!aliveRef.current) return;
-          const pollMsg =
-            pollErr instanceof Error ? pollErr.message : 'Could not refresh experiment';
-          devWarn('ExperimentDetailScreen', `detail poll failed — ${experimentId.slice(0, 8)}… — ${pollMsg}`);
-          setError('Could not refresh experiment — transient network or server error.');
-        }
-      }, DETAIL_POLL_MS);
-    },
-    [experimentId, stopDetailPoll],
-  );
-
-  useEffect(() => {
-    aliveRef.current = true;
-    const abortHydrate = new AbortController();
-
-    const stall = createStallWatcher({
-      scope: 'ExperimentDetailScreen',
-      operation: 'detail hydrate',
-      alive: () => aliveRef.current,
-      afterMs: LOADING_STALL_AFTER_MS,
-      repeatMs: LOADING_STALL_REPEAT_MS,
-      onWarning: (text) => setLoadFeed((f) => appendFeedEntry(f, text, 'warning')),
-    });
-
-    const applyProg: ExperimentProgressCallback = (u: FetchProgressUpdate) => {
-      if (!aliveRef.current) return;
-      if (u.type === 'downloading') {
-        setReceivedBytes(u.receivedBytes);
-        setTotalBytes(u.totalBytes);
-        return;
-      }
-      setLoadFeed((f) =>
-        appendFeedEntry(f, u.text, u.variant === 'warning' ? 'warning' : 'default'),
-      );
-    };
-
-    async function hydrate() {
-      const hasSeed = initialExperiment?.experiment_id === experimentId;
-      devInfo(
-        'ExperimentDetailScreen',
-        hasSeed
-          ? `hydrate started — refreshing ${experimentId.slice(0, 8)}… (seed from list)`
-          : `hydrate started — loading ${experimentId.slice(0, 8)}…`,
-      );
-      setHydrating(true);
-      setError(null);
-      if (!hasSeed) {
-        setDetail(null);
-      }
-      setLoadFeed([
-        {
-          id: 'h0',
-          text: hasSeed
-            ? 'Refreshing run rows and live status…'
-            : 'Fetching experiment and run rows…',
-          variant: 'default',
-        },
-      ]);
-      setReceivedBytes(null);
-      setTotalBytes(null);
-      stall.start();
-
-      stopDetailPoll();
-
-      let loadedStatus: ExperimentStatus | undefined;
-      let loadedCompletedAt: string | null | undefined;
-
-      try {
-        const loaded = hasSeed
-          ? await getExperiment(experimentId, abortHydrate.signal)
-          : await getExperimentWithProgress(experimentId, applyProg, abortHydrate.signal);
-        stall.stop();
-        if (!aliveRef.current) return;
-        loadedStatus = loaded.status;
-        loadedCompletedAt = loaded.completed_at;
-        setDetail(loaded as unknown as ExperimentDetail);
-        const runs = (loaded as { runs?: unknown[] }).runs;
-        const runRows = Array.isArray(runs) ? runs.length : 0;
-        devInfo(
-          'ExperimentDetailScreen',
-          `hydrate OK — ${experimentId.slice(0, 8)}… status=${loaded.status}, ${runRows} run row(s)`,
-        );
-        setLoadFeed((f) =>
-          appendFeedEntry(
-            f,
-            isRunningExperimentStatus(loaded.status, loadedCompletedAt)
-              ? 'Run rows loaded — live polling while experiment is running.'
-              : 'Run rows loaded.',
-            'default',
-          ),
-        );
-      } catch (err) {
-        stall.stop();
-        if (!aliveRef.current) return;
-        if (err instanceof DOMException && err.name === 'AbortError') return;
-        const msg =
-          err instanceof Error ? err.message : 'Failed to load experiment';
-        devWarn('ExperimentDetailScreen', `hydrate failed — ${experimentId.slice(0, 8)}… — ${msg}`);
-        setError(msg);
-        setLoadFeed((f) => appendFeedEntry(f, `Failed: ${msg}`, 'warning'));
-      } finally {
-        stall.stop();
-        if (aliveRef.current) {
-          setHydrating(false);
-          startDetailPollIfRunning(loadedStatus, loadedCompletedAt);
-        }
-      }
-    }
-
-    void hydrate();
-
-    return () => {
-      aliveRef.current = false;
-      abortHydrate.abort();
-      stall.stop();
-      stopDetailPoll();
-    };
-  }, [experimentId, initialExperiment, startDetailPollIfRunning, stopDetailPoll]);
-
-  const refreshDetailAfterControl = useCallback(async () => {
-    const refreshed = await getExperiment(experimentId);
-    setDetail(refreshed as unknown as ExperimentDetail);
-    const refreshedCompletedAt = refreshed.completed_at;
-    if (isRunningExperimentStatus(refreshed.status, refreshedCompletedAt)) {
-      startDetailPollIfRunning(refreshed.status, refreshedCompletedAt);
-    } else {
-      stopDetailPoll();
-    }
-    setError(null);
-  }, [experimentId, startDetailPollIfRunning, stopDetailPoll]);
-
-  async function handleDelete() {
-    setDeleting(true);
-    try {
-      await deleteExperiment(experimentId);
-      devInfo('ExperimentDetailScreen', `delete OK — experiment ${experimentId.slice(0, 8)}…`);
-      setShowDeleteModal(false);
-      onBack();
-    } catch (err) {
-      const msg = err instanceof Error ? err.message : 'Failed to delete experiment';
-      devWarn('ExperimentDetailScreen', `delete failed — ${experimentId.slice(0, 8)}… — ${msg}`);
-      setError(msg);
-      setShowDeleteModal(false);
-    } finally {
-      setDeleting(false);
-    }
-  }
+  const {
+    detail,
+    error,
+    setError,
+    hydrating,
+    showDeleteModal,
+    setShowDeleteModal,
+    deleting,
+    loadFeed,
+    receivedBytes,
+    totalBytes,
+    dbStats,
+    dbStatsLoading,
+    runsCurrentPage,
+    setRunsCurrentPage,
+    runsItemsPerPage,
+    handleRunsItemsPerPageChange,
+    refreshDetailAfterControl,
+    handleDelete,
+  } = useExperimentDetail({
+    experimentId,
+    initialExperiment,
+    initialDbStats,
+    onDeleted: onBack,
+  });
 
   const isTerminal = detail && isTerminalExperimentStatus(detail.status, detail.completed_at);
   const isRunning = isRunningExperimentStatus(detail?.status, detail?.completed_at);
