+++ frontend/src/components/experimentDetail/index.ts
@@ -1,0 +1,9 @@
+export { detailIcons } from './icons';
+export type { ExperimentDetail } from './types';
+export { PHASE_ORDER } from './types';
+export { default as PhaseIndicator } from './PhaseIndicator';
+export { default as ProgressSubtitle } from './ProgressSubtitle';
+export { default as MetadataItem } from './MetadataItem';
+export { default as StatusBadge } from './StatusBadge';
+export { default as StatCard } from './StatCard';
+export { default as DimensionBadge } from './DimensionBadge';
