+++ frontend/src/components/experimentDetailProgress.ts
@@ -44,3 +44,39 @@ export function formatTimeWithUnits(totalSeconds: number): string {
   }
   return `${minutes}m ${seconds}s`;
 }
+
+export function formatDuration(startedAt?: string, completedAt?: string | null): string {
+  if (!startedAt || !completedAt) return '—';
+  const ms = new Date(completedAt).getTime() - new Date(startedAt).getTime();
+  if (ms < 1000) return `${ms}ms`;
+  return formatTimeWithUnits(ms / 1000);
+}
+
+export function parseSafeTimestamp(value: string | undefined): number | null {
+  if (!value) return null;
+  const parsed = new Date(value).getTime();
+  if (Number.isNaN(parsed)) return null;
+  return parsed;
+}
+
+export function formatDurationFromRuns(
+  runs: { created_at?: string; updated_at?: string }[] = [],
+  startedAt?: string,
+  completedAt?: string | null,
+): string {
+  const runStartedAt = Math.min(
+    ...runs.map((run) => parseSafeTimestamp(run.created_at)).filter((value): value is number => value !== null),
+  );
+  const runCompletedAt = Math.max(
+    ...runs.map((run) => parseSafeTimestamp(run.updated_at)).filter((value): value is number => value !== null),
+  );
+
+  const allRunsHaveTimestamps =
+    Number.isFinite(runStartedAt) && Number.isFinite(runCompletedAt) && runCompletedAt >= runStartedAt;
+
+  if (allRunsHaveTimestamps) {
+    return formatDuration(new Date(runStartedAt).toISOString(), new Date(runCompletedAt).toISOString());
+  }
+
+  return formatDuration(startedAt, completedAt);
+}
